# openapi.yaml
openapi: 3.0.3
info:
  title: distrebute.com Platform API
  version: 0.1.0
  description: |
    Unified API surface for distrebute.com — the 22 UI templates wired to
    backend services. Every endpoint is owned by exactly one microservice
    (see `x-owner` extension). Gateway routes path prefixes to services.

    Auth model: short-lived JWT access tokens (15 min) + opaque refresh
    tokens via the auth-service. Passkey / WebAuthn primary, password fallback.
  contact:
    name: distrebute.com platform team
    email: api@distrebute.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local docker-compose gateway
  - url: https://api.distrebute.com/v1
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /auth/sso/{provider}:
    post:
      tags: [auth]
      summary: SSO sign-in (apple, google, facebook)
      x-owner: auth-service
      x-templates: [18]
      security: []
      parameters:
        - in: path
          name: provider
          required: true
          schema: { type: string, enum: [apple, google, facebook] }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                idToken: { type: string }
                authCode: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
  /auth/email/start:
    post:
      tags: [auth]
      summary: Send magic link / OTP to email
      x-owner: auth-service
      x-templates: [18]
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '202': { description: Code sent }
  /auth/refresh:
    post:
      tags: [auth]
      summary: Exchange refresh token for a new access token
      x-owner: auth-service
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
...
  /auth/2fa/verify:
    post:
      tags: [auth]
      summary: Verify TOTP code, enable 2FA
      x-owner: auth-service
      x-templates: [18]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                code: { type: string, example: "847291" }
      responses:
        '200': { description: 2FA enabled }
...
  /auth/events:
    get:
      tags: [auth]
      summary: Recent sign-in events (success, blocked, suspicious)
      x-owner: auth-service
      x-templates: [18, 20]
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SecurityEvent' }

  # ---- USERS, PROFILES, DEVICES, KIDS --------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TokenPair:
      type: object
      properties:
        accessToken: { type: string, description: "JWT, 15-min TTL" }
        refreshToken: { type: string, description: "opaque, 30-day TTL" }
        expiresIn: { type: integer, example: 900 }
...
    SecurityEvent:
      type: object
      properties:
        id: { type: string }
        type: { type: string, enum: [signin, signout, password_change, blocked_attempt, device_added] }
        device: { type: string }
        location: { type: string }
        ipMasked: { type: string }
        at: { type: string, format: date-time }
        flagged: { type: boolean }
...